Use floats again

+++ ros1/src/data/march_realsense_reader/include/pointcloud_processor/region_creator.h
@@ -84,7 +84,7 @@ private:
         const boost::shared_ptr<RegionVector> last_region_vector,
         const PointCloud::Ptr last_pointcloud,
         const Normals::Ptr last_pointcloud_normals,
-        const double& last_tolerance);
+        const float& last_tolerance);
 
     // Splits a region Vector into the regions considered too large, just right,
     // and too small
@@ -97,7 +97,7 @@ private:
     // Creates a potential region vector from a pointcloud with a certain
     // tolerance
     bool getRegionVectorFromTolerance(const PointCloud::Ptr pointcloud,
-        const Normals::Ptr pointcloud_normals, const double& tolerance,
+        const Normals::Ptr pointcloud_normals, const float& tolerance,
         boost::shared_ptr<RegionVector> region_vector);
 
     // Add the right regions to the region points and region normals vectors
@@ -117,7 +117,7 @@ private:
     // Process the invalid regions with the new tolerance
     // This method makes a call to this method if the invalid region is large
     // enough
-    bool processInvalidRegions(const double& next_tolerance,
+    bool processInvalidRegions(const float& next_tolerance,
         const PointCloud::Ptr invalid_pointcloud,
         Normals::Ptr invalid_pointcloud_normals,
         const boost::shared_ptr<RegionVector> invalid_regions,
@@ -136,13 +136,13 @@ private:
     int max_valid_cluster_size;
     int min_desired_cluster_size;
     int max_desired_cluster_size;
-    double smoothness_threshold;
-    double smoothness_threshold_lower_bound;
-    double smoothness_threshold_upper_bound;
+    float smoothness_threshold;
+    float smoothness_threshold_lower_bound;
+    float smoothness_threshold_upper_bound;
     float curvature_threshold;
     bool use_recursive_growing;
-    double tolerance_change_factor_decrease;
-    double tolerance_change_factor_increase;
+    float tolerance_change_factor_decrease;
+    float tolerance_change_factor_increase;
     boost::shared_ptr<RegionVector> region_vector_;
     int number_of_recursive_calls;
 };
